File structure typo fix
