+++ CalCount/src/main/java/com/revature/controllers/UserController.java
@@ -1,4 +1,4 @@
-package com.reavture.controllers;
+package com.revature.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
@@ -7,23 +7,23 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.reavture.models.Meal;
-import com.revature.services.MealService;
+import com.revature.models.User;
+import com.revature.services.UserService;
 
 @RestController
-@RequestMapping("/meals")
-public class MealController {
+@RequestMapping("/users")
+public class UserController {
 	
 	@Autowired
-	private MealService mealService;
+	private UserService userService;
 	
 	@PostMapping(path = "/new", consumes = MediaType.APPLICATION_JSON_VALUE)
-	public void saveMeal(@RequestBody Meal meal) {
-		mealService.saveMeal(meal);
+	public void saveMeal(@RequestBody User user) {
+		userService.saveMeal(user);
 	}
 	
-	public Meal findById(int Id) {
-		return mealService.findById(Id);
+	public boolean Login(String email, String Password) {
+		userService.findUserByEmail(email);
+		return true;
 	}
-
 }
