+++ CalCount/src/main/java/com/revature/models/User.java
@@ -1,4 +1,4 @@
-package com.reavture.models;
+package com.revature.models;
 
 import java.util.List;
