+++ CalCount/src/main/java/com/revature/repositories/UserRepository.java
@@ -2,7 +2,7 @@ package com.revature.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.reavture.models.User;
+import com.revature.models.User;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
