+++ CalCount/src/main/java/com/revature/services/MealService.java
@@ -3,7 +3,7 @@ package com.revature.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.reavture.models.Meal;
+import com.revature.models.Meal;
 import com.revature.repositories.MealRepository;
