+++ CalCount/src/main/java/com/revature/services/UserService.java
@@ -3,7 +3,7 @@ package com.revature.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.reavture.models.User;
+import com.revature.models.User;
 import com.revature.repositories.UserRepository;
 
 @Service
